correction cpp

+++ cpp06/ex00/includes/Para.hpp
@@ -6,7 +6,7 @@
 /*   By: mbonnet <mbonnet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/10 17:11:35 by mbonnet           #+#    #+#             */
-/*   Updated: 2022/03/15 11:52:26 by mbonnet          ###   ########.fr       */
+/*   Updated: 2022/03/28 22:20:43 by mbonnet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,8 +34,8 @@ class Para
 		std::string	_inff;
 		int			_type;
 		double		_Double;
-		float		_Float;
-		int			_Int;
+		double		_Float;
+		double		_Int;
 		char		_Char;
 		std::string	_val;
 };
